+++ Assets/Packages/com.studio23.ss2.upmassistant/Editor/Data/PackageDependency.cs
@@ -1,4 +1,4 @@
-﻿namespace Studio23.SS2.UPMAssistant.Editor
+﻿namespace Studio23.SS2.UPMAssistant.Editor.Data
 {
     public class PackageDependency
     {