+++ Assets/Packages/com.studio23.ss2.upmassistant/Editor/Data/ScopedRegistry.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Studio23.SS2.UPMAssistant.Editor
+namespace Studio23.SS2.UPMAssistant.Editor.Data
 {
     
     public class ScopedRegistry