+++ Assets/Packages/com.studio23.ss2.upmassistant/Editor/Data/UPMAssistantData.cs
@@ -1,0 +1,17 @@
+using System;
+
+namespace Studio23.SS2.UPMAssistant.Editor.Data
+{
+    [System.Serializable]
+
+    public class PackageNameData
+    {
+        public string PackageName;
+    }
+
+    [Serializable]
+    public class LicenseURLData
+    {
+        public string LicenseURL;
+    }
+}