+++ Assets/Packages/com.studio23.ss2.upmassistant/CHANGELOG.md
@@ -10,3 +10,13 @@
 - dependency add and remove bug fixed.
 - refresh and delete button add.
 - tests asmdef included.
+
+
+## [1.0.3](2023-10-19)
+
+- Delete system, delete folder structure with save data
+- Create folder structure if not exist
+- Automatically update pacakge.json name, company name, package version while updating in Player Settings
+- Automatically get Licenses from online, fetch, show in dropdown, download, save, update, delete, update package.josn while updating license
+- Markup edit system
+- Added assembly under Runtime folder, added two folders under Runtime folder: Core, Data